fix: update Day.schema to use tinyint for weekdays instead of int

+++ models/rahel-hirsch/database/schemas/Day.schema.ts
@@ -1,12 +1,12 @@
-import { mysqlTable, int } from 'drizzle-orm/mysql-core';
+import { mysqlTable, tinyint, int } from 'drizzle-orm/mysql-core';
 
 export const Days_Schema = mysqlTable('Day', {
     DayId: int('DayId').primaryKey().autoincrement(),
-    monday: int('monday').notNull(),
-    tuesday: int('tuesday').notNull(),
-    wednesday: int('wednesday').notNull(),
-    thursday: int('thursday').notNull(),
-    friday: int('friday').notNull(),
+    monday: tinyint('monday').notNull(),
+    tuesday: tinyint('tuesday').notNull(),
+    wednesday: tinyint('wednesday').notNull(),
+    thursday: tinyint('thursday').notNull(),
+    friday: tinyint('friday').notNull(),
     places: int('places').notNull(),
     takenPlaces: int('takenPlaces')
 });
